pyproject.toml
[tool.poetry]
name = "vikuatools"
version = "0.1.0"
description = "Tools for BI and Analytics aimed to shape smart cities"
authors = ["Rodrigo Serrano"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = ">=1.3.5"
google-cloud-bigquery = ">=2.32.0"
pyarrow = ">=6.0.1"
DateTime = ">=4.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
